refactor(web): extract backtesting result data
